Support `arg.func` wrapped queryset function in multi-object actions
One can declare `queryset` attribute on actions with a custom function, wrapped with `arg.func` from the `python-args` library. This works fine for single-object actions, as the `get_queryset()` method is overriden to support loading of those "lazy" functions.

This breaks if the action is used for multiple objects, as the queryset is handed over to Django for iteration, but it is not iterable, but a `arg.func` object.

Type: bug